# specs/001-zero-training-mvp/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: LiverRa v1 MVP HTTP API
  version: 1.0.0
  description: |
    Public HTTP API for the LiverRa zero-training liver AI pipeline platform.

    **Auth**: OIDC Bearer token from AWS Cognito (research A.1) with `custom:tenant_id` claim.
    **RBAC**: Every protected op declares `x-required-permission`. Step-up auth (≤5 min freshness) is marked `x-step-up: true`.
    **Errors**: RFC 7807 Problem Details. `type`, `title`, `status`, `detail`, plus LiverRa extensions `tenant_id`, `audit_event_id`, `claim_key` (never PHI).
    **Audit**: Every state-changing op emits a FHIR AuditEvent (chain-of-hashes research A.3).
servers:
  - url: https://app.liverra.ai/api/v1
    description: Production
  - url: http://localhost:7050/api/v1
    description: Local dev (FastAPI ml-inference)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuditEventSummary:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        actor: { type: string }
        subject: { type: string }
        timestamp: { type: string, format: date-time }
        outcome: { type: string, enum: [success, denied, error] }
        chain_sequence_no: { type: integer }
    ClaimRegistryEntry:
      type: object
      properties:
        claim_key: { type: string, enum: [parenchyma_volumetry, flr, couinaud_segmentation, vessel_identification, lesion_detection, lesion_classification, surgical_planning] }
        status: { type: string, enum: [ruo, under_conformity_assessment, cleared] }
        effective_from: { type: string, format: date-time }
        regulatory_reference: { type: string, nullable: true }
...
paths:
...
  /compliance/audit-summary:
    get:
      tags: [compliance]
      x-required-permission: compliance.generate_audit_summary
      parameters:
        - { name: tenant_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Tamper-evident audit summary with chain verification
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: '#/components/schemas/AuditEventSummary' } }
                  chain_valid: { type: boolean }
                  chain_first_invalid_sequence_no: { type: integer, nullable: true }
                  merkle_root_for_window: { type: string, description: "hex" }
                  s3_anchor_uris: { type: array, items: { type: string, format: uri } }
  /compliance/ruo-spot-check:
    post:
      tags: [compliance]
      x-required-permission: compliance.spot_check_ruo
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sample_size: { type: integer, default: 20 }
      responses:
        '200':
          description: Sampled artifacts with watermark-region highlighted
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    artifact_url: { type: string, format: uri }
                    watermark_bbox: { type: array, items: { type: integer } }
                    pass: { type: boolean, nullable: true, description: "Reviewer fills post-check" }
  /compliance/claim-registry:
    get:
      tags: [compliance]
      x-required-permission: compliance.view_mbom
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ClaimRegistryEntry' }
    put:
      tags: [compliance]
      x-required-permission: compliance.toggle_claim_registry
      x-step-up: true
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimRegistryEntry' }
      responses:
        '200': { description: Updated (subsequent exports narrow disclaimer scope) }
...